+++ ktlint/src/main/kotlin/com/pinterest/ktlint/internal/KtlintCommandLine.kt
@@ -116,15 +116,17 @@ internal class KtlintCommandLine {
 
     @Option(
         names = ["--debug"],
-        description = ["Turn on debug output"],
+        description = ["Turn on debug output. Deprecated, use '--log-level=debug' instead."],
     )
-    var debug: Boolean = false
+    @Deprecated(message = "Replaced with minLogLevel")
+    var debugOld: Boolean? = null
 
     @Option(
         names = ["--trace"],
-        description = ["Turn on trace output"],
+        description = ["Turn on trace output. Deprecated, use '--log-level=trace' instead."],
     )
-    var trace: Boolean = false
+    @Deprecated(message = "Replaced with minLogLevel")
+    var trace: Boolean? = null
 
     @Option(
         names = ["--disabled_rules"],
@@ -194,9 +196,10 @@ internal class KtlintCommandLine {
 
     @Option(
         names = ["--verbose", "-v"],
-        description = ["Show error codes"],
+        description = ["Show error codes. Deprecated, use '--log-level=info' instead."],
     )
-    private var verbose: Boolean = false
+    @Deprecated(message = "Replaced with minLogLevel")
+    private var verbose: Boolean? = null
 
     @Option(
         names = ["--editorconfig"],
@@ -224,14 +227,32 @@ internal class KtlintCommandLine {
     @Parameters(hidden = true)
     private var patterns = ArrayList<String>()
 
+    @Option(
+        names = ["--log-level", "-l"],
+        description = ["Defines the minimum log level (trace, debug, info, warn, error) or none to suppress all logging"],
+        converter = [LogLevelConverter::class],
+    )
+    private var minLogLevel: Level = Level.INFO
+
     private val tripped = AtomicBoolean()
     private val fileNumber = AtomicInteger()
     private val errorNumber = AtomicInteger()
 
+    internal var debug: Boolean = false
+        get() = Level.DEBUG.isGreaterOrEqual(minLogLevel)
+        private set
+
     fun run() {
-        if (verbose) {
-            debug = true
+        if (debugOld != null || trace != null || verbose != null) {
+            if (minLogLevel == Level.OFF) {
+                minLogLevel = Level.ERROR
+            }
+            configureLogger().error {
+                "Options '--debug', '--trace', '--verbose' and '-v' are deprecated and replaced with option '--log-level=<level>' or '-l=<level>'."
+            }
+            exitProcess(1)
         }
+
         logger = configureLogger()
 
         assertStdinAndPatternsFromStdinOptionsMutuallyExclusive()
@@ -311,11 +332,7 @@ internal class KtlintCommandLine {
         KotlinLogging
             .logger {}
             .setDefaultLoggerModifier { logger ->
-                (logger.underlyingLogger as Logger).level = when {
-                    trace -> Level.TRACE
-                    debug -> Level.DEBUG
-                    else -> Level.INFO
-                }
+                (logger.underlyingLogger as Logger).level = minLogLevel
             }
             .initKtLintKLogger()
 
@@ -471,7 +488,6 @@ internal class KtlintCommandLine {
                     reporterId,
                     split.lastOrNull { it.startsWith("artifact=") }?.let { it.split("=")[1] },
                     mapOf(
-                        "verbose" to verbose.toString(),
                         "color" to color.toString(),
                         "color_name" to colorName,
                         "format" to format.toString(),
@@ -541,7 +557,7 @@ internal class KtlintCommandLine {
                     "",
                     "Internal Error (${e.ruleId}) in file '$filename' at position '${e.line}:${e.col}. " +
                         "Please create a ticket at https://github.com/pinterest/ktlint/issues " +
-                        "(if possible, please re-run with the --debug flag to get the stacktrace " +
+                        "(if possible, please re-run with the --log-level=debug flag to get the stacktrace " +
                         "and provide the source code that triggered an error)",
                 )
             }
@@ -675,4 +691,18 @@ internal class KtlintCommandLine {
         val config: Map<String, String>,
         val output: String?,
     )
+}
+
+private class LogLevelConverter : CommandLine.ITypeConverter<Level> {
+    @Throws(Exception::class)
+    override fun convert(value: String?): Level =
+        when (value?.uppercase()) {
+            "TRACE" -> Level.TRACE
+            "DEBUG" -> Level.DEBUG
+            "INFO" -> Level.INFO
+            "WARN" -> Level.WARN
+            "ERROR" -> Level.ERROR
+            "NONE" -> Level.OFF
+            else -> Level.INFO
+        }
 }